2-Langchain Projects/Rag document Q&A Chatbot/requirements.txt
langchain==0.1.16
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
frontend
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_groq
uvicorn
fastapi
langserve
duckdb
pandas
openai
langchain_chroma
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtuber_transcript_api
unstructured
pytube
numexpr